backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
pydantic-settings==2.10.1
supabase==2.7.4
python-dateutil==2.9.0.post0
scikit-learn==1.7.2
joblib==1.5.2
numpy==2.3.3
torch==2.8.0
sentence-transformers==5.1.1
redis==6.4.0
